requirements.txt
datasets
pillow
streamlit